LayoutTests/performance-api/paint-timing/contentful-with-background-image-expected.txt

PASS entryList.getEntries().length is 1
PASS entryList.getEntriesByName("first-contentful-paint").length is 1
PASS entryList.getEntriesByType("paint").length is 1
PASS didReceiveFirstContentfulPaint is false
PASS fcpEntry.name is "first-contentful-paint"
PASS fcpEntry.entryType is "paint"
PASS didReceiveFirstContentfulPaint is true
PASS performance.getEntries().length is 2
PASS performance.getEntriesByName("first-contentful-paint").length is 1
PASS performance.getEntriesByName("first-contentful-paint", "paint").length is 1
PASS performance.getEntriesByType("paint").length is 1
PASS PerformanceObserver first-contentful-paint callback wasn't fired again
PASS successfullyParsed is true

TEST COMPLETE
